Move checkUai job from oneShot folder
